requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
django-heroku==0.3.1